Added a check before deleting a category if it can be deleted

+++ src/main/java/me/ezzedine/mohammed/personalspace/category/core/ArticleCategoryService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -31,16 +32,23 @@ public class ArticleCategoryService implements ArticleCategoryFetcher, ArticleCa
     }
 
     @Override
-    public void delete(String id) throws ArticleCategoryNotFoundException {
-        validateIdExists(id);
+    public void delete(String id) throws ArticleCategoryNotFoundException, CannotDeleteArticleCategoryException {
+        ArticleCategory articleCategory = validateCategoryExists(id);
+
+        if (!articleCategory.canBeDeleted()) {
+            throw new CannotDeleteArticleCategoryException(id);
+        }
 
         storage.delete(id);
     }
 
-    private void validateIdExists(String id) throws ArticleCategoryNotFoundException {
-        if (!storage.categoryExists(id)) {
+    private ArticleCategory validateCategoryExists(String id) throws ArticleCategoryNotFoundException {
+        Optional<ArticleCategory> optionalCategory = storage.fetch(id);
+        if (optionalCategory.isEmpty()) {
             throw new ArticleCategoryNotFoundException(id);
         }
+
+        return optionalCategory.get();
     }
 
     private void validateIdIsNotTaken(PersistArticleCategoryRequest request, String categoryId) throws ArticleCategoryIdAlreadyExistsException {
